+++ NEWS.md
@@ -11,6 +11,7 @@ Removed:
 Added:
 
 * View classes (`Nanoc::ItemView` etc)
+* `DataSource#new_item`
 
 ## 3.8.0 (2015-05-04)